Migrate ChipsInput to Incubator ChipsInput (#2483)
* Migrate ChipsInput to Incubator ChipsInput

* removed old ChipsInput componenet

* Moved Incubator ChipsInput to components, fixed review notes
