+++ src/components/chipsInput/assets/xSmall.png
