+++ src/components/chipsInput/assets/xSmall@1.5x.png
