+++ src/components/chipsInput/assets/xSmall@2x.png
